+++ zeno/zeno/nodes/prim/PrimitiveClip.cpp
@@ -1,6 +1,7 @@
 #include <zeno/zeno.h>
 #include <zeno/types/PrimitiveObject.h>
 #include <zeno/types/NumericObject.h>
+#include <zeno/utils/prim_ops.h>
 #include <unordered_map>
 #include <stdexcept>
